Close 9/16 follow-ups: qty carryover, whole-order CTP, integer sale qty.
Unfinished boards return to a PMC confirm pool instead of auto-reschedule;
create-order CTP shares one capacity pool and stays advisory; sale units stay whole numbers.

+++ apps/api/routers/mis.py
@@ -23,7 +23,7 @@ def _role(x_demo_role: str | None) -> str:
 
 class MisOrderLineIn(BaseModel):
     item_code: str
-    qty: float = Field(gt=0)
+    qty: int = Field(gt=0)
     unit: str = "BOX"
     unit_price: float = Field(default=0, ge=0)
